Finished insertion form.
Finished "Inserir reparos" form. Any repair inserted will be registered in the database, and any new clients added will be accessible to the user to use again. Added a failsafe that prevents the same client from being added twice, and the repair won't go foward unless a client is added.